MOVES THE SCHEMATICS TO THE APPROPRIATE FILE
